Seperated Database service
