+++ lambda/package.json
@@ -1,6 +1,6 @@
 {
   "name": "lambda",
-  "version": "1.0.0",
+  "version": "0.1.0",
   "main": "bot.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
